initial work on synchrony plugin. Do not deploy, code is not in a complete state
moving some declarations around

adding in the form stuff for the customer side and hopefully fetching a token

update the fields of the payment form on the checkout page.

moving some filter outside the class and adding validation methods

adding in failure static variables and properly testing the client id token for validity

doing some stuff to prep for an interim pay here page

adding the redirect to checkout pay for order page with cancel button to submit to Synchrony

moving the extra synchrony fields out of checkout page and into the special pay now page.

also adding in method for the response urls